rename read.ts to their correct class names
