+++ typescript/sdk/src/core/EvmCoreReader.ts
@@ -4,8 +4,8 @@ import { Mailbox__factory } from '@hyperlane-xyz/core';
 import { Address, objMap, promiseObjAll } from '@hyperlane-xyz/utils';
 
 import { DEFAULT_CONTRACT_READ_CONCURRENCY } from '../consts/concurrency.js';
-import { EvmHookReader } from '../hook/read.js';
-import { EvmIsmReader } from '../ism/read.js';
+import { EvmHookReader } from '../hook/EvmHookReader.js';
+import { EvmIsmReader } from '../ism/EvmIsmReader.js';
 import { MultiProvider } from '../providers/MultiProvider.js';
 import { ChainNameOrId } from '../types.js';
